requirements.txt
# PyTorch
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0

# HuggingFace
transformers>=4.38.2
tokenizers>=0.15.2
huggingface-hub>=0.21.4
datasets>=2.18.0
accelerate>=0.27.2

# Vision
timm>=0.9.16
opencv-python-headless>=4.9.0.80
grad-cam>=1.5.2
pillow>=10.3.0

# FastAPI
fastapi>=0.110.0
uvicorn[standard]>=0.27.1
python-multipart>=0.0.9
httpx>=0.27.0
starlette>=0.36.3

# Task queue
celery>=5.3.6
redis>=5.0.3
kombu>=5.3.4

# Supabase
supabase>=2.4.1
postgrest>=0.16.1
gotrue>=2.4.0
storage3>=0.7.0
realtime>=1.0.2

# LangChain
langchain>=0.1.16
langchain-core>=0.1.46
langchain-community>=0.0.38
langchain-anthropic>=0.1.11
anthropic>=0.23.1
tavily-python>=0.3.3

# Gradio
gradio>=4.26.0

# ML utilities
scikit-learn>=1.4.1
numpy>=1.26.4
matplotlib>=3.8.3

# Audio
soundfile>=0.12.1
librosa>=0.10.1
audioread>=3.0.1

# Utilities
python-dotenv>=1.0.1
loguru>=0.7.2
slowapi>=0.1.9
limits>=3.10.1
pydantic>=2.6.4
pydantic-settings>=2.2.1

# Dev / test
pytest>=8.1.1
pytest-asyncio>=0.23.6
pytest-cov>=5.0.0
black>=24.3.0
ruff>=0.3.4
